refactor(Manager): Rename BaseExecutor
BREAKING CHANGE: Package exports `Manager` instead of `BaseExecutor`.

This renaming is in preparaton for encapsulating `Manager`'s functionality into several separate classes.
