+++ src/browser/init.ts
@@ -1,7 +1,7 @@
 import '@stencila/components'
 import { getLogger } from '@stencila/logga'
 import { CodeChunk, softwareSession, SoftwareSession } from '@stencila/schema'
-import { BaseExecutor } from '../base/BaseExecutor'
+import { Manager } from '../base/Manager'
 import { HttpAddressInitializer } from '../base/Transports'
 import { discover } from '../http/discover'
 import { HttpClient } from '../http/HttpClient'
@@ -9,7 +9,7 @@ import { WebSocketClient } from '../ws/WebSocketClient'
 
 const log = getLogger('executa:browser')
 
-let executor: BaseExecutor
+let executor: Manager
 let session: null | SoftwareSession = null
 
 const executeCodeChunk = async (codeChunk: CodeChunk): Promise<CodeChunk> => {
@@ -37,7 +37,7 @@ const onReadyHandler = (): void => {
 }
 
 export const init = (address: HttpAddressInitializer): void => {
-  executor = new BaseExecutor(
+  executor = new Manager(
     [() => discover(address)],
     [HttpClient, WebSocketClient]
   )
